+++ tests/test_unit/test_napari_plugin/test_poses_loader_widget.py
@@ -10,7 +10,7 @@ from napari.settings import get_settings
 from pytest import DATA_PATHS
 from qtpy.QtWidgets import QComboBox, QDoubleSpinBox, QLineEdit, QPushButton
 
-from movement.napari._loader_widgets import PosesLoader
+from movement.napari.widgets import PosesLoader
 
 
 # ------------------- tests for widget instantiation--------------------------#
